Removed _base...was doing nothing and was wrong - wrong classes in message Deleted trait _Base and use Base to extend all, to provide messages. static::class sufficient to get correct class names in messages
Moved all Entity static data and methods to an EntityManager class, which objects are tracked/managed by an EntityManagerManager class - using statics for global access.

Removed the Entity::init approach for Model creation/loading. Had a apotential for cycles.  Was causing warnings anyway because of different signature to Entity::init which was caputing the entity and model classes, and the porperty information
Property information now coming from the Model.
Now the EM guesses the Model Class from the Entity::class - which is sufficient to connect the two.
The Entity:constructor requests an EM object using the static::class, and the EMM returns the correct EM object.
All Entities are recorded in their EM object - yet to decide how to manage auto-persist etc.
May decide to add some Model to the EM as a plugin, rather tha  hardwire it.
Need overrides and extra mapping info

Must test a custom method to provide the EM/Model details...Entity::registerEntitymanager called from EM/EMM if it exists on $entityClass.

Also must test the use of Entity assignments to Entity properties, and mapping to ID's etc...